Move knapsack example to new example folder
