+++ examples/__init__.py
@@ -1,0 +1,1 @@
+"""Examples showcasing how to use optiframe."""
